requirements.txt
pandas
torch
numpy
numba
geopandas
rasterio
scikit-learn
pillow
transformers
torchvision
segmentation_models_pytorch